Fixed js file stuff
